debug

+++ src/diffusers/models/embeddings.py
@@ -725,7 +725,7 @@ class SizeEmbedder(nn.Module):
         size = size.reshape(-1)
         size_freq = get_timestep_embedding(
             size, self.frequency_embedding_size, downscale_freq_shift=0, flip_sin_to_cos=True
-        )
+        ).to(size.dtype)
 
         size_emb = self.mlp(size_freq)
         size_emb = size_emb.reshape(current_batch_size, dims * self.outdim)
